requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
torch
pandas
numpy
scikit-learn
